requirements.txt
matplotlib==3.6.2
pandas==1.4.1
psycopg2_binary==2.9.5
scipy==1.9.3
seaborn==0.11.2
SQLAlchemy==1.4.41
streamlit==1.20.0
